+++ resources/js/app.js
@@ -1,1 +1,1 @@
-require('./bootstrap');
+// require('./bootstrap');
