+++ resources/views/components/layouts/header.blade.php
@@ -5,8 +5,7 @@
     <title>@yield('title')</title>
     <link href="{{ asset('css/app.css') }}" rel="stylesheet">
     <link href="{{ mix('css/custom.css') }}" rel="stylesheet">
-    <link href="{{ mix('css/calendar.css') }}" rel="stylesheet">
-    <script src="{{ mix('js/calendar.js') }}"></script>
+    <script src="{{ mix('js/custom.js') }}"></script>
     {{-- <script src="https://cdn.jsdelivr.net/npm/fullcalendar-scheduler@5.3.0/main.js"></script> --}}
 </head>
 @endsection
