+++ webpack.mix.js
@@ -12,13 +12,16 @@ const mix = require('laravel-mix');
  */
 
 mix
-    .js('resources/js/app.js', 'public/js')
-    .js('resources/js/custom.js', 'public/js')
-    .js('resources/js/calendar.js', 'public/js')
+    .js('resources/js/schedule/components/calendar.js', 'public/js/custom.js')
+    .js('resources/js/schedule/app.js', 'public/js/custom.js')
+    .js('resources/js/vital/components/graph.js', 'public/js/custom.js')
+    .js('resources/js/vital/components/tab.js', 'public/js/custom.js')
+    .js('resources/js/vital/app.js', 'public/js/custom.js')
     .postCss('resources/css/app.css', 'public/css', [
         require("tailwindcss"),
     ])
-    .sass('resources/scss/calendar.scss', 'public/css')
-    .sass('resources/scss/custom.scss', 'public/css');
+    .sass('resources/scss/calendar/app.scss', 'public/css/custom.css')
+    .sass('resources/scss/vital/app.scss', 'public/css/custom.css')
+    .sass('resources/scss/schedule/app.scss', 'public/css/custom.css');
 
 ;
